@@ -34,35 +34,45 @@
         "source": "https://github.com/phpList/core"
     },
     "require": {
-        "php": "^7.2|^8.0",
-        "doctrine/orm": "^2.5.0",
-        "doctrine/common": "^2.6.0",
-        "doctrine/doctrine-bundle": "^1.8.0",
-        "symfony/symfony": "^3.4.37",
-        "symfony/monolog-bundle": "^3.1.0",
-        "symfony/dependency-injection": "^3.4.37",
-        "symfony/config": "^3.4.37",
-        "symfony/yaml": "^3.4.37",
-        "jms/serializer-bundle": "^3.8.0",
-        "sensio/framework-extra-bundle": "^5.1.0",
-        "sensio/distribution-bundle": "^5.0.6"
+        "php": "^8.1",
+        "symfony/dependency-injection": "^6.4",
+        "symfony/config": "^6.4",
+        "symfony/yaml": "^6.4",
+        "symfony/error-handler": "^6.4",
+        "symfony/serializer": "^6.4",
+        "symfony/monolog-bundle": "^3.10",
+        "symfony/serializer-pack": "^1.3",
+        "symfony/orm-pack": "^2.4",
+        "doctrine/orm": "^3.3",
+        "doctrine/cache": "^2.2",
+        "symfony/asset": "^6.4",
+        "symfony/security-csrf": "^6.4",
+        "symfony/form": "^6.4",
+        "symfony/validator": "^6.4",
+        "doctrine/doctrine-fixtures-bundle": "^3.7",
+        "doctrine/instantiator": "^2.0",
+        "masterminds/html5": "^2.9"
     },
     "require-dev": {
-        "phpunit/phpunit": "^6.5.6",
-        "phpunit/phpunit-mock-objects": "^5.0.6",
-        "phpunit/dbunit": "^3.0.0",
+        "phpunit/phpunit": "^9.5",
         "guzzlehttp/guzzle": "^6.3.0",
         "squizlabs/php_codesniffer": "^3.2.0",
-        "phpstan/phpstan": "^0.7.0|0.12.57",
-        "nette/caching": "^2.5.0|^3.0.0",
-        "nikic/php-parser": "^3.1.0",
+        "phpstan/phpstan": "^1.10",
+        "nette/caching": "^3.0.0",
+        "nikic/php-parser": "^4.19.1",
         "phpmd/phpmd": "^2.6.0",
-        "composer/composer": "^1.6.0",
-        "doctrine/instantiator": "^1.0.5"
+        "symfony/test-pack": "^1.1",
+        "symfony/process": "^6.4",
+        "composer/composer": "^2.7",
+        "symfony/framework-bundle": "^6.4",
+        "symfony/http-kernel": "^6.4",
+        "symfony/http-foundation": "^6.4",
+        "symfony/routing": "^6.4",
+        "symfony/console": "^6.4"
     },
     "suggest": {
-        "phplist/web-frontend": "4.0.x-dev",
-        "phplist/rest-api": "4.0.x-dev"
+        "phplist/web-frontend": "5.0.x-dev",
+        "phplist/rest-api": "5.0.x-dev"
     },
     "autoload": {
         "psr-4": {
