@@ -1,15 +1,16 @@
 <?php
+
 declare(strict_types=1);
 
 namespace PhpList\Core\Domain\Model\Messaging;
 
+use DateTime;
 use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\Common\Collections\Collection;
-use Doctrine\ORM\Mapping;
-use Doctrine\ORM\Mapping\Column;
-use Doctrine\ORM\Proxy\Proxy;
-use JMS\Serializer\Annotation\ExclusionPolicy;
-use JMS\Serializer\Annotation\Expose;
+use Doctrine\ORM\Mapping as ORM;
+use PhpList\Core\Domain\Model\Subscription\Subscription;
+use Symfony\Component\Serializer\Annotation\Ignore;
+use Symfony\Component\Serializer\Annotation\SerializedName;
 use PhpList\Core\Domain\Model\Identity\Administrator;
 use PhpList\Core\Domain\Model\Interfaces\CreationDate;
 use PhpList\Core\Domain\Model\Interfaces\DomainModel;
@@ -18,278 +19,188 @@ use PhpList\Core\Domain\Model\Interfaces\ModificationDate;
 use PhpList\Core\Domain\Model\Traits\CreationDateTrait;
 use PhpList\Core\Domain\Model\Traits\IdentityTrait;
 use PhpList\Core\Domain\Model\Traits\ModificationDateTrait;
+use Symfony\Component\Serializer\Attribute\Groups;
+use Symfony\Component\Serializer\Attribute\MaxDepth;
 
 /**
  * This class represents an administrator who can log to the system, is allowed to administer
  * selected lists (as the owner), send campaigns to these lists and edit subscribers.
- *
- * @Mapping\Entity(repositoryClass="PhpList\Core\Domain\Repository\Messaging\SubscriberListRepository")
- * @Mapping\Table(name="phplist_list")
- * @Mapping\HasLifecycleCallbacks
- * @ExclusionPolicy("all")
- *
  * @author Oliver Klee <oliver@phplist.com>
  */
+#[ORM\Entity(repositoryClass: 'PhpList\Core\Domain\Repository\Messaging\SubscriberListRepository')]
+#[ORM\Table(name: 'phplist_list')]
+#[ORM\Index(name: 'nameidx', columns: ['name'])]
+#[ORM\Index(name: 'listorderidx', columns: ['listorder'])]
+#[ORM\HasLifecycleCallbacks]
 class SubscriberList implements DomainModel, Identity, CreationDate, ModificationDate
 {
     use IdentityTrait;
     use CreationDateTrait;
     use ModificationDateTrait;
 
-    /**
-     * @var string
-     * @Column
-     * @Expose
-     */
-    private $name = '';
+    #[ORM\Column]
+    #[SerializedName('name')]
+    #[Groups(['SubscriberList'])]
+    private string $name = '';
 
-    /**
-     * @var string
-     * @Column
-     * @Expose
-     */
-    private $description = '';
+    #[ORM\Column]
+    #[SerializedName('description')]
+    #[Groups(['SubscriberList'])]
+    private string $description = '';
 
-    /**
-     * @var \DateTime|null
-     * @Column(type="datetime", nullable=true, name="entered")
-     * @Expose
-     */
-    protected $creationDate = null;
+    #[ORM\Column(name: 'entered', type: 'datetime', nullable: true)]
+    #[SerializedName('creation_date')]
+    #[Groups(['SubscriberList'])]
+    protected ?DateTime $creationDate = null;
 
-    /**
-     * @var \DateTime|null
-     * @Column(type="datetime", name="modified")
-     */
-    protected $modificationDate = null;
+    #[ORM\Column(name: 'modified', type: 'datetime')]
+    #[Ignore]
+    protected ?DateTime $modificationDate = null;
 
-    /**
-     * @var int
-     * @Column(type="integer", name="listorder")
-     * @Expose
-     */
-    private $listPosition = 0;
+    #[ORM\Column(name: 'listorder', type: 'integer')]
+    #[SerializedName('list_position')]
+    #[Groups(['SubscriberList'])]
+    private ?int $listPosition;
 
-    /**
-     * @var string
-     * @Column(name="prefix")
-     * @Expose
-     */
-    private $subjectPrefix = '';
+    #[ORM\Column(name: 'prefix')]
+    #[SerializedName('subject_prefix')]
+    #[Groups(['SubscriberList'])]
+    private ?string $subjectPrefix;
 
-    /**
-     * @var bool
-     * @Column(type="boolean", name="active")
-     * @Expose
-     */
-    private $public = false;
+    #[ORM\Column(name: 'active', type: 'boolean')]
+    #[SerializedName('public')]
+    #[Groups(['SubscriberList'])]
+    private bool $public;
 
-    /**
-     * @var string
-     * @Column
-     * @Expose
-     */
-    private $category = '';
+    #[ORM\Column]
+    #[SerializedName('category')]
+    #[Groups(['SubscriberList'])]
+    private string $category;
 
-    /**
-     * @var Administrator
-     * @Mapping\ManyToOne(targetEntity="PhpList\Core\Domain\Model\Identity\Administrator")
-     * @Mapping\JoinColumn(name="owner")
-     */
-    private $owner = null;
+    #[ORM\ManyToOne(targetEntity: 'PhpList\Core\Domain\Model\Identity\Administrator')]
+    #[ORM\JoinColumn(name: 'owner')]
+    #[Ignore]
+    private ?Administrator $owner = null;
 
-    /**
-     * @var Collection
-     * @Mapping\OneToMany(
-     *     targetEntity="PhpList\Core\Domain\Model\Subscription\Subscription",
-     *     mappedBy="subscriberList",
-     *     cascade={"remove"}
-     * )
-     */
-    private $subscriptions = null;
+    #[ORM\OneToMany(
+        targetEntity: 'PhpList\Core\Domain\Model\Subscription\Subscription',
+        mappedBy: 'subscriberList',
+        cascade: ['remove'],
+        orphanRemoval: true,
+    )]
+    #[MaxDepth(1)]
+    private Collection $subscriptions;
 
-    /**
-     * @var Collection
-     * @Mapping\ManyToMany(
-     *     targetEntity="PhpList\Core\Domain\Model\Subscription\Subscriber",
-     *     inversedBy="subscribedLists",
-     *     fetch="EXTRA_LAZY"
-     * )
-     * @Mapping\JoinTable(name="phplist_listuser",
-     *     joinColumns={@Mapping\JoinColumn(name="listid")},
-     *     inverseJoinColumns={@Mapping\JoinColumn(name="userid")}
-     * )
-     */
-    private $subscribers = null;
-
-    /**
-     * The constructor.
-     */
     public function __construct()
     {
         $this->subscriptions = new ArrayCollection();
-        $this->subscribers = new ArrayCollection();
+        $this->listPosition = 0;
+        $this->subjectPrefix = '';
+        $this->category = '';
+        $this->public = false;
     }
 
-    /**
-     * @return string
-     */
     public function getName(): string
     {
         return $this->name;
     }
 
-    /**
-     * @param string $name
-     *
-     * @return void
-     */
-    public function setName(string $name)
+    public function setName(string $name): void
     {
         $this->name = $name;
     }
 
-    /**
-     * @return string
-     */
     public function getDescription(): string
     {
         return $this->description;
     }
 
-    /**
-     * @param string $description
-     *
-     * @return void
-     */
-    public function setDescription(string $description)
+    public function setDescription(string $description): void
     {
         $this->description = $description;
     }
 
-    /**
-     * @return int
-     */
     public function getListPosition(): int
     {
         return $this->listPosition;
     }
 
-    /**
-     * @param int $listPosition
-     *
-     * @return void
-     */
-    public function setListPosition(int $listPosition)
+    public function setListPosition(int $listPosition): void
     {
         $this->listPosition = $listPosition;
     }
 
-    /**
-     * @return string
-     */
     public function getSubjectPrefix(): string
     {
         return $this->subjectPrefix;
     }
 
-    /**
-     * @param string $subjectPrefix
-     *
-     * @return void
-     */
-    public function setSubjectPrefix(string $subjectPrefix)
+    public function setSubjectPrefix(string $subjectPrefix): void
     {
         $this->subjectPrefix = $subjectPrefix;
     }
 
-    /**
-     * @return bool
-     */
     public function isPublic(): bool
     {
-        return $this->public;
+        return $this->public ?? false;
     }
 
-    /**
-     * @param bool $public
-     *
-     * @return void
-     */
-    public function setPublic(bool $public)
+    public function setPublic(bool $public): void
     {
         $this->public = $public;
     }
 
-    /**
-     * @return string
-     */
     public function getCategory(): string
     {
         return $this->category;
     }
 
-    /**
-     * @param string $category
-     *
-     * @return void
-     */
-    public function setCategory(string $category)
+    public function setCategory(string $category): void
     {
         $this->category = $category;
     }
 
-    /**
-     * @return Administrator|Proxy|null
-     */
-    public function getOwner()
+    public function getOwner(): ?Administrator
     {
         return $this->owner;
     }
 
-    /**
-     * @param Administrator $owner
-     *
-     * @return void
-     */
-    public function setOwner(Administrator $owner)
+    public function setOwner(Administrator $owner): void
     {
         $this->owner = $owner;
     }
 
-    /**
-     * @return Collection
-     */
     public function getSubscriptions(): Collection
     {
         return $this->subscriptions;
     }
 
-    /**
-     * @param Collection $subscriptions
-     *
-     * @return void
-     */
-    public function setSubscriptions(Collection $subscriptions)
+    public function addSubscription(Subscription $subscription): self
     {
-        $this->subscriptions = $subscriptions;
+        if (!$this->subscriptions->contains($subscription)) {
+            $this->subscriptions->add($subscription);
+            $subscription->setSubscriberList($this);
+        }
+
+        return $this;
     }
 
-    /**
-     * @return Collection
-     */
+    public function removeSubscription(Subscription $subscription): self
+    {
+        if ($this->subscriptions->removeElement($subscription)) {
+            $subscription->setSubscriberList(null);
+        }
+
+        return $this;
+    }
+
     public function getSubscribers(): Collection
     {
-        return $this->subscribers;
-    }
+        $result = new ArrayCollection();
+        foreach ($this->subscriptions as $subscription) {
+            $result->add($subscription->getSubscriber());
+        }
 
-    /**
-     * @param Collection $subscribers
-     *
-     * @return void
-     */
-    public function setSubscribers(Collection $subscribers)
-    {
-        $this->subscribers = $subscribers;
+        return $result;
     }
 }
