@@ -1,13 +1,14 @@
 <?php
+
 declare(strict_types=1);
 
 namespace PhpList\Core\Domain\Model\Subscription;
 
-use Doctrine\ORM\Mapping;
-use Doctrine\ORM\Mapping\Column;
-use Doctrine\ORM\Proxy\Proxy;
-use JMS\Serializer\Annotation\ExclusionPolicy;
-use JMS\Serializer\Annotation\Expose;
+use DateTime;
+use Doctrine\ORM\Mapping as ORM;
+use Doctrine\Persistence\Proxy;
+use Symfony\Component\Serializer\Annotation\Ignore;
+use Symfony\Component\Serializer\Annotation\SerializedName;
 use PhpList\Core\Domain\Model\Interfaces\CreationDate;
 use PhpList\Core\Domain\Model\Interfaces\DomainModel;
 use PhpList\Core\Domain\Model\Interfaces\ModificationDate;
@@ -16,88 +17,64 @@ use PhpList\Core\Domain\Model\Traits\CreationDateTrait;
 use PhpList\Core\Domain\Model\Traits\ModificationDateTrait;
 
 /**
- * This class represents asubscriber who can subscribe to multiple subscriber lists and can receive email messages from
+ * This class represents subscriber who can subscribe to multiple subscriber lists and can receive email messages from
  * campaigns for those subscriber lists.
- *
- * @Mapping\Entity(repositoryClass="PhpList\Core\Domain\Repository\Subscription\SubscriptionRepository")
- * @Mapping\Table(name="phplist_listuser")
- * @Mapping\HasLifecycleCallbacks
- * @ExclusionPolicy("all")
- *
  * @author Oliver Klee <oliver@phplist.com>
  */
+#[ORM\Entity(repositoryClass: 'PhpList\Core\Domain\Repository\Subscription\SubscriptionRepository')]
+#[ORM\Table(name: 'phplist_listuser')]
+#[ORM\Index(name: 'userenteredidx', columns: ['userid', 'entered'])]
+#[ORM\Index(name: 'userlistenteredidx', columns: ['userid', 'entered', 'listid'])]
+#[ORM\Index(name: 'useridx', columns: ['userid'])]
+#[ORM\Index(name: 'listidx', columns: ['listid'])]
+#[ORM\HasLifecycleCallbacks]
 class Subscription implements DomainModel, CreationDate, ModificationDate
 {
     use CreationDateTrait;
     use ModificationDateTrait;
 
-    /**
-     * @var \DateTime|null
-     * @Column(type="datetime", nullable=true, name="entered")
-     * @Expose
-     */
-    protected $creationDate = null;
+    #[ORM\Column(name: 'entered', type: 'datetime', nullable: true)]
+    #[SerializedName('creation_date')]
+    protected ?DateTime $creationDate = null;
 
-    /**
-     * @var \DateTime|null
-     * @Column(type="datetime", name="modified")
-     */
-    protected $modificationDate = null;
+    #[ORM\Column(name: 'modified', type: 'datetime')]
+    #[Ignore]
+    protected ?DateTime $modificationDate = null;
 
-    /**
-     * @var Subscriber|Proxy|null
-     * @Mapping\Id
-     * @Mapping\ManyToOne(
-     *     targetEntity="PhpList\Core\Domain\Model\Subscription\Subscriber",
-     *     inversedBy="subscriptions"
-     * )
-     * @Mapping\JoinColumn(name="userid")
-     */
-    private $subscriber = null;
+    #[ORM\Id]
+    #[ORM\ManyToOne(
+        targetEntity: 'PhpList\Core\Domain\Model\Subscription\Subscriber',
+        inversedBy: 'subscriptions'
+    )]
+    #[ORM\JoinColumn(name: 'userid')]
+    #[SerializedName('subscriber')]
+    private ?Subscriber $subscriber = null;
 
-    /**
-     * @var SubscriberList|Proxy|null
-     * @Mapping\Id
-     * @Mapping\ManyToOne(
-     *     targetEntity="PhpList\Core\Domain\Model\Messaging\SubscriberList",
-     *     inversedBy="subscriptions"
-     * )
-     * @Mapping\JoinColumn(name="listid")
-     */
-    private $subscriberList = null;
+    #[ORM\Id]
+    #[ORM\ManyToOne(
+        targetEntity: 'PhpList\Core\Domain\Model\Messaging\SubscriberList',
+        inversedBy: 'subscriptions'
+    )]
+    #[ORM\JoinColumn(name: 'listid', onDelete: 'CASCADE')]
+    #[Ignore]
+    private ?SubscriberList $subscriberList = null;
 
-    /**
-     * @return Subscriber|Proxy|null
-     */
-    public function getSubscriber()
+    public function getSubscriber(): Subscriber|Proxy|null
     {
         return $this->subscriber;
     }
 
-    /**
-     * @param Subscriber $subscriber
-     *
-     * @return void
-     */
-    public function setSubscriber(Subscriber $subscriber)
+    public function setSubscriber(?Subscriber $subscriber): void
     {
         $this->subscriber = $subscriber;
     }
 
-    /**
-     * @return SubscriberList|Proxy|null
-     */
-    public function getSubscriberList()
+    public function getSubscriberList(): ?SubscriberList
     {
         return $this->subscriberList;
     }
 
-    /**
-     * @param SubscriberList $subscriberList
-     *
-     * @return void
-     */
-    public function setSubscriberList(SubscriberList $subscriberList)
+    public function setSubscriberList(?SubscriberList $subscriberList): void
     {
         $this->subscriberList = $subscriberList;
     }
