@@ -1,8 +1,11 @@
 <?php
+
 declare(strict_types=1);
 
 namespace PhpList\Core\Domain\Repository\Identity;
 
+use Doctrine\ORM\EntityManagerInterface;
+use Doctrine\ORM\Mapping\ClassMetadata;
 use PhpList\Core\Domain\Model\Identity\Administrator;
 use PhpList\Core\Domain\Repository\AbstractRepository;
 use PhpList\Core\Security\HashGenerator;
@@ -14,34 +17,29 @@ use PhpList\Core\Security\HashGenerator;
  */
 class AdministratorRepository extends AbstractRepository
 {
-    /**
-     * @var HashGenerator
-     */
-    private $hashGenerator = null;
+    private HashGenerator $hashGenerator;
 
-    /**
-     * @param HashGenerator $hashGenerator
-     * @required
-     *
-     * @return void
-     */
-    public function injectHashGenerator(HashGenerator $hashGenerator)
-    {
-        $this->hashGenerator = $hashGenerator;
+    public function __construct(
+        EntityManagerInterface $entityManager,
+        ClassMetadata $class,
+        HashGenerator $hashGenerator = null
+    ) {
+        parent::__construct($entityManager, $class);
+        $this->hashGenerator = $hashGenerator ?? new HashGenerator();
     }
 
     /**
      * Finds the Administrator with the given login credentials. Returns null if there is no match,
      * i.e., if the login credentials are incorrect.
      *
-     * This also checks that the administrator is a super user.
+     * This also checks that the administrator is a superuser.
      *
      * @param string $loginName
      * @param string $plainTextPassword
      *
      * @return Administrator|null
      */
-    public function findOneByLoginCredentials(string $loginName, string $plainTextPassword)
+    public function findOneByLoginCredentials(string $loginName, string $plainTextPassword): ?Administrator
     {
         $passwordHash = $this->hashGenerator->createPasswordHash($plainTextPassword);
 
