@@ -1,19 +1,19 @@
 <?php
+
 declare(strict_types=1);
 
 namespace PhpList\Core\Tests\Integration\Routing;
 
-use Doctrine\Common\Annotations\SimpleAnnotationReader;
 use PhpList\Core\Core\ApplicationKernel;
 use PhpList\Core\Core\ApplicationStructure;
 use PhpList\Core\Core\Bootstrap;
 use PhpList\Core\Core\Environment;
 use PhpList\Core\Routing\ExtraLoader;
 use PHPUnit\Framework\TestCase;
-use Sensio\Bundle\FrameworkExtraBundle\Routing\AnnotatedRouteControllerLoader;
+use Symfony\Bundle\FrameworkBundle\Routing\AttributeRouteControllerLoader;
+use Symfony\Component\Config\FileLocator;
 use Symfony\Component\Config\Loader\LoaderResolver;
-use Symfony\Component\HttpKernel\Config\FileLocator;
-use Symfony\Component\Routing\Loader\AnnotationDirectoryLoader;
+use Symfony\Component\Routing\Loader\AttributeDirectoryLoader;
 use Symfony\Component\Routing\Loader\YamlFileLoader;
 use Symfony\Component\Routing\RouteCollection;
 
@@ -24,33 +24,30 @@ use Symfony\Component\Routing\RouteCollection;
  */
 class ExtraLoaderTest extends TestCase
 {
-    /**
-     * @var ExtraLoader
-     */
-    private $subject = null;
+    private ?ExtraLoader $subject = null;
+    private ?ApplicationKernel $kernel = null;
 
-    /**
-     * @var ApplicationKernel
-     */
-    private $kernel = null;
-
-    protected function setUp()
+    protected function setUp(): void
     {
         $bootstrap = Bootstrap::getInstance();
         $bootstrap->setEnvironment(Environment::TESTING)->configure();
 
         $this->kernel = $bootstrap->getApplicationKernel();
         $this->kernel->boot();
-        $container = $this->kernel->getContainer();
 
-        /** @var FileLocator $locator */
-        $locator = $container->get('file_locator');
-        $routeControllerLoader = new AnnotatedRouteControllerLoader(new SimpleAnnotationReader());
+        $locator = new FileLocator([
+            $this->kernel->getProjectDir() . '/src/EmptyStartPageBundle/Controller',
+            $this->kernel->getProjectDir() . '/src/EmptyStartPageBundle',
+            $this->kernel->getProjectDir() . '/config',
+        ]);
+
+        $attributeLoader = new AttributeRouteControllerLoader();
+        $attributeDirectoryLoader = new AttributeDirectoryLoader($locator, $attributeLoader);
 
         $loaderResolver = new LoaderResolver(
             [
                 new YamlFileLoader($locator),
-                new AnnotationDirectoryLoader($locator, $routeControllerLoader),
+                $attributeDirectoryLoader,
             ]
         );
 
@@ -58,17 +55,17 @@ class ExtraLoaderTest extends TestCase
         $this->subject->setResolver($loaderResolver);
     }
 
-    protected function tearDown()
+    protected function tearDown(): void
     {
         $this->kernel->shutdown();
         Bootstrap::purgeInstance();
     }
 
-    /**
-     * @test
-     */
-    public function loadReturnsRouteCollection()
+    public function loadReturnsRouteCollection(): void
     {
-        static::assertInstanceOf(RouteCollection::class, $this->subject->load('', 'extra'));
+        $routeCollection = $this->subject->load('@EmptyStartPageBundle/Controller/', 'extra');
+
+        self::assertInstanceOf(RouteCollection::class, $routeCollection);
+        self::assertNotNull($routeCollection->get('empty_start_page'));
     }
 }
