@@ -1,4 +1,5 @@
 <?php
+
 declare(strict_types=1);
 
 namespace PhpList\Core\Tests\System\ApplicationBundle;
@@ -16,57 +17,32 @@ class PhpListApplicationBundleTest extends TestCase
 {
     use SymfonyServerTrait;
 
-    /**
-     * @var Client
-     */
-    private $httpClient = null;
+    private ?Client $httpClient = null;
 
-    protected function setUp()
+    protected function setUp(): void
     {
+        parent::setUp();
         $this->httpClient = new Client(['http_errors' => false]);
     }
 
-    protected function tearDown()
+    protected function tearDown(): void
     {
         $this->stopSymfonyServer();
+        $this->httpClient = null;
+        parent::tearDown();
     }
 
-    /**
-     * @return string[][]
-     */
-    public function environmentDataProvider(): array
+    public function testHomepageReturnsSuccess(): void
     {
-        return [
-            'test' => ['test'],
-            'dev' => ['dev'],
-        ];
-    }
+        $this->startSymfonyServer();
+        $response = $this->httpClient->get('/api/v2', [
+            'base_uri' => $this->getBaseUrl(),
+        ]);
 
-    /**
-     * @test
-     * @param string $environment
-     * @dataProvider environmentDataProvider
-     */
-    public function homepageReturnsSuccess(string $environment)
-    {
-        $this->startSymfonyServer($environment);
-
-        $response = $this->httpClient->get('/', ['base_uri' => $this->getBaseUrl()]);
-
-        static::assertSame(200, $response->getStatusCode());
-    }
-
-    /**
-     * @test
-     * @param string $environment
-     * @dataProvider environmentDataProvider
-     */
-    public function homepageReturnsDummyContent(string $environment)
-    {
-        $this->startSymfonyServer($environment);
-
-        $response = $this->httpClient->get('/', ['base_uri' => $this->getBaseUrl()]);
-
-        static::assertContains('This page has been intentionally left empty.', $response->getBody()->getContents());
+        self::assertSame(200, $response->getStatusCode());
+        self::assertStringContainsString(
+            'This page has been intentionally left empty.',
+            $response->getBody()->getContents()
+        );
     }
 }
