@@ -1,9 +1,10 @@
 <?php
+
 declare(strict_types=1);
 
 namespace PhpList\Core\Tests\Unit\EmptyStartPageBundle;
 
-use PhpList\Core\EmptyStartPageBundle\PhpListEmptyStartPageBundle;
+use PhpList\Core\EmptyStartPageBundle\EmptyStartPageBundle;
 use PHPUnit\Framework\TestCase;
 use Symfony\Component\HttpKernel\Bundle\Bundle;
 
@@ -14,21 +15,15 @@ use Symfony\Component\HttpKernel\Bundle\Bundle;
  */
 class PhpListEmptyStartPageBundleTest extends TestCase
 {
-    /**
-     * @var PhpListEmptyStartPageBundle
-     */
-    private $subject = null;
+    private EmptyStartPageBundle $subject;
 
-    protected function setUp()
+    protected function setUp(): void
     {
-        $this->subject = new PhpListEmptyStartPageBundle();
+        $this->subject = new EmptyStartPageBundle();
     }
 
-    /**
-     * @test
-     */
-    public function classIsBundle()
+    public function testClassIsBundle(): void
     {
-        static::assertInstanceOf(Bundle::class, $this->subject);
+        self::assertInstanceOf(Bundle::class, $this->subject);
     }
 }
