@@ -1,7 +1,10 @@
 <?php
+
 declare(strict_types=1);
 
 namespace PhpList\Core\Core;
+
+use RuntimeException;
 
 /**
  * This class provides information about the current application and its file structure.
@@ -15,7 +18,7 @@ class ApplicationStructure
      *
      * @return string the absolute path without the trailing slash.
      *
-     * @throws \RuntimeException if there is no composer.json in the application root
+     * @throws RuntimeException if there is no composer.json in the application root
      */
     public function getCorePackageRoot(): string
     {
@@ -33,7 +36,7 @@ class ApplicationStructure
      *
      * @return string the absolute path without the trailing slash
      *
-     * @throws \RuntimeException if there is no composer.json in the application root
+     * @throws RuntimeException if there is no composer.json in the application root
      */
     public function getApplicationRoot(): string
     {
@@ -48,7 +51,7 @@ class ApplicationStructure
         }
 
         if (!file_exists($applicationRoot . '/composer.json')) {
-            throw new \RuntimeException(
+            throw new RuntimeException(
                 'There is no composer.json in the supposed application root "' . $applicationRoot . '".',
                 1501169001
             );
