@@ -1,50 +1,38 @@
 <?php
+
 declare(strict_types=1);
 
 namespace PhpList\Core\Domain\Model\Traits;
 
-use Doctrine\ORM\Mapping;
+use DateTime;
+use Doctrine\ORM\Mapping as ORM;
 
 /**
  * This trait provides an automatic modification date for models.
  *
  * This is the default implementation of the ModificationDate interface.
  *
- * Please note that this trait requires the model to have the "HasLifecycleCallbacks" annotation in the class doc block,
+ * Please note that this trait requires the model to have the "HasLifecycleCallbacks" attribute in the class doc block,
  * and also to have a $modificationDate property with the correct column name mapping.
  *
  * @author Oliver Klee <oliver@phplist.com>
  */
 trait ModificationDateTrait
 {
-    /**
-     * @return \DateTime|null
-     */
-    public function getModificationDate()
+    public function getModificationDate(): ?DateTime
     {
         return $this->modificationDate;
     }
 
-    /**
-     * @param \DateTime $modificationDate
-     *
-     * @return void
-     */
-    private function setModificationDate(\DateTime $modificationDate)
+    private function setModificationDate(DateTime $modificationDate): void
     {
         $this->modificationDate = $modificationDate;
     }
 
-    /**
-     * Updates the modification date to now.
-     *
-     * @Mapping\PrePersist
-     * @Mapping\PreUpdate
-     *
-     * @return void
-     */
-    public function updateModificationDate()
+    #[ORM\PrePersist]
+    #[ORM\PreUpdate]
+    public function updateModificationDate(): void
     {
-        $this->setModificationDate(new \DateTime());
+        $this->setModificationDate(new DateTime());
     }
 }
