@@ -1,29 +1,28 @@
 <?php
+
 declare(strict_types=1);
 
 namespace PhpList\Core\EmptyStartPageBundle\Controller;
 
-use Sensio\Bundle\FrameworkExtraBundle\Configuration\Method;
-use Sensio\Bundle\FrameworkExtraBundle\Configuration\Route;
-use Symfony\Bundle\FrameworkBundle\Controller\Controller;
+use InvalidArgumentException;
+use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\Routing\Annotation\Route;
 
 /**
  * This controller provides an empty start page.
  *
  * @author Oliver Klee <oliver@phplist.com>
  */
-class DefaultController extends Controller
+class DefaultController extends AbstractController
 {
     /**
-     * @Route("/")
-     * @Method("GET")
+     * An empty start page route.
      *
-     * @return Response
-     *
-     * @throws \InvalidArgumentException
+     * @throws InvalidArgumentException
      */
-    public function indexAction(): Response
+    #[Route('/api/v2', name: 'empty_start_page', methods: ['GET'])]
+    public function index(): Response
     {
         return new Response('This page has been intentionally left empty.');
     }
