@@ -1,4 +1,5 @@
 <?php
+
 declare(strict_types=1);
 
 namespace PhpList\Core\EmptyStartPageBundle;
@@ -10,6 +11,6 @@ use Symfony\Component\HttpKernel\Bundle\Bundle;
  *
  * @author Oliver Klee <oliver@phplist.com>
  */
-class PhpListEmptyStartPageBundle extends Bundle
+class EmptyStartPageBundle extends Bundle
 {
 }
