@@ -1,12 +1,16 @@
 <?php
+
 declare(strict_types=1);
 
 namespace PhpList\Core\Tests\Integration\Security;
 
+use Doctrine\ORM\Tools\SchemaTool;
 use PhpList\Core\Domain\Model\Identity\Administrator;
 use PhpList\Core\Security\Authentication;
 use PhpList\Core\TestingSupport\Traits\DatabaseTestTrait;
-use PHPUnit\Framework\TestCase;
+use PhpList\Core\Tests\Integration\Domain\Repository\Fixtures\AdministratorFixture;
+use PhpList\Core\Tests\Integration\Domain\Repository\Fixtures\AdministratorTokenWithAdministratorFixture;
+use Symfony\Bundle\FrameworkBundle\Test\KernelTestCase;
 use Symfony\Component\HttpFoundation\Request;
 
 /**
@@ -14,62 +18,40 @@ use Symfony\Component\HttpFoundation\Request;
  *
  * @author Oliver Klee <oliver@phplist.com>
  */
-class AuthenticationTest extends TestCase
+class AuthenticationTest extends KernelTestCase
 {
     use DatabaseTestTrait;
 
-    /**
-     * @var string
-     */
-    const ADMINISTRATOR_TABLE_NAME = 'phplist_admin';
+    private ?Authentication $subject = null;
 
-    /**
-     * @var string
-     */
-    const TOKEN_TABLE_NAME = 'phplist_admintoken';
-
-    /**
-     * @var Authentication
-     */
-    private $subject = null;
-
-    protected function setUp()
+    protected function setUp(): void
     {
-        $this->setUpDatabaseTest();
+        parent::setUp();
+        $this->loadSchema();
 
-        $this->subject = $this->container->get(Authentication::class);
+        $this->subject = self::getContainer()->get(Authentication::class);
     }
 
-    /**
-     * @test
-     */
-    public function subjectIsAvailableViaContainer()
+    protected function tearDown(): void
     {
-        static::assertInstanceOf(Authentication::class, $this->subject);
+        $schemaTool = new SchemaTool($this->entityManager);
+        $schemaTool->dropDatabase();
+        parent::tearDown();
     }
 
-    /**
-     * @test
-     */
-    public function classIsRegisteredAsSingletonInContainer()
+    public function testSubjectIsAvailableViaContainer()
     {
-        static::assertSame($this->subject, $this->container->get(Authentication::class));
+        self::assertInstanceOf(Authentication::class, $this->subject);
     }
 
-    /**
-     * @test
-     */
-    public function authenticateByApiKeyWithValidApiKeyReturnsMatchingAdministrator()
+    public function testClassIsRegisteredAsSingletonInContainer()
     {
-        $this->getDataSet()->addTable(
-            static::ADMINISTRATOR_TABLE_NAME,
-            __DIR__ . '/../Domain/Repository/Fixtures/Administrator.csv'
-        );
-        $this->getDataSet()->addTable(
-            static::TOKEN_TABLE_NAME,
-            __DIR__ . '/../Domain/Repository/Fixtures/AdministratorTokenWithAdministrator.csv'
-        );
-        $this->applyDatabaseChanges();
+        self::assertSame($this->subject, self::getContainer()->get(Authentication::class));
+    }
+
+    public function testAuthenticateByApiKeyWithValidApiKeyReturnsMatchingAdministrator()
+    {
+        $this->loadFixtures([AdministratorFixture::class, AdministratorTokenWithAdministratorFixture::class]);
 
         $apiKey = 'cfdf64eecbbf336628b0f3071adba762';
         $request = new Request();
@@ -77,24 +59,13 @@ class AuthenticationTest extends TestCase
 
         $result = $this->subject->authenticateByApiKey($request);
 
-        static::assertInstanceOf(Administrator::class, $result);
-        static::assertSame(1, $result->getId());
+        self::assertInstanceOf(Administrator::class, $result);
+        self::assertSame(1, $result->getId());
     }
 
-    /**
-     * @test
-     */
-    public function authenticateByApiKeyWithValidApiKeyAndDisabledAdministratorReturnsNull()
+    public function testAuthenticateByApiKeyWithValidApiKeyAndDisabledAdministratorReturnsNull()
     {
-        $this->getDataSet()->addTable(
-            static::ADMINISTRATOR_TABLE_NAME,
-            __DIR__ . '/../Domain/Repository/Fixtures/Administrator.csv'
-        );
-        $this->getDataSet()->addTable(
-            static::TOKEN_TABLE_NAME,
-            __DIR__ . '/../Domain/Repository/Fixtures/AdministratorTokenWithAdministrator.csv'
-        );
-        $this->applyDatabaseChanges();
+        $this->loadFixtures([AdministratorFixture::class, AdministratorTokenWithAdministratorFixture::class]);
 
         $apiKey = 'cfdf64eecbbf336628b0f3071adba765';
         $request = new Request();
@@ -102,19 +73,12 @@ class AuthenticationTest extends TestCase
 
         $result = $this->subject->authenticateByApiKey($request);
 
-        static::assertNull($result);
+        self::assertNull($result);
     }
 
-    /**
-     * @test
-     */
-    public function authenticateByApiKeyWithValidApiKeyForInexistentAdministratorReturnsNull()
+    public function testAuthenticateByApiKeyWithValidApiKeyForInexistentAdministratorReturnsNull()
     {
-        $this->getDataSet()->addTable(
-            static::TOKEN_TABLE_NAME,
-            __DIR__ . '/../Domain/Repository/Fixtures/AdministratorTokenWithAdministrator.csv'
-        );
-        $this->applyDatabaseChanges();
+        $this->loadFixtures([AdministratorFixture::class, AdministratorTokenWithAdministratorFixture::class]);
 
         $apiKey = 'cfdf64eecbbf336628b0f3071adba763';
         $request = new Request();
@@ -125,16 +89,9 @@ class AuthenticationTest extends TestCase
         static::assertNull($result);
     }
 
-    /**
-     * @test
-     */
-    public function authenticateByApiKeyWithValidApiKeyForNonSuperUserAdministratorReturnsNull()
+    public function testAuthenticateByApiKeyWithValidApiKeyForNonSuperUserAdministratorReturnsNull()
     {
-        $this->getDataSet()->addTable(
-            static::TOKEN_TABLE_NAME,
-            __DIR__ . '/../Domain/Repository/Fixtures/AdministratorTokenWithAdministrator.csv'
-        );
-        $this->applyDatabaseChanges();
+        $this->loadFixtures([AdministratorFixture::class, AdministratorTokenWithAdministratorFixture::class]);
 
         $apiKey = 'cfdf64eecbbf336628b0f3071adba764';
         $request = new Request();
@@ -142,6 +99,6 @@ class AuthenticationTest extends TestCase
 
         $result = $this->subject->authenticateByApiKey($request);
 
-        static::assertNull($result);
+        self::assertNull($result);
     }
 }
